+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <!-- ## [Unreleased] -->
-<!-- ### Breaking changes ⚠>
+<!-- ### Breaking changes ⚠ -->
 <!-- ### Added ✔️-->
 <!-- ### Changed 🛠️-->
 <!-- ### Deprecated 🛑-->
@@ -12,6 +12,7 @@
 ## [Unreleased]
 ### Added ✔️
 * **Multi tenant**: Allow to configure tenants using the preferences.
+* **DefaultSecurityAutoConfiguration**: Allow to use SSL in LDAP connections.
 ### Fixed 🐛
 * **ODMSAutoConfigure**: Allows the dmsColumnHelper to get the nameConvention from the application.
 ## [3.14.0] - 2024-02-23